chore: Move scala debug related types to debugger
